# Revert raw event consumer system and traits optimization