--- a/hikari/impl/shard.py
+++ b/hikari/impl/shard.py
@@ -70,7 +70,7 @@ _T: typing.Final[str] = sys.intern("t")
 _S: typing.Final[str] = sys.intern("s")
 _OP: typing.Final[str] = sys.intern("op")
 
-# Opcodes.
+# Opcodes
 _DISPATCH: typing.Final[int] = 0
 _HEARTBEAT: typing.Final[int] = 1
 _IDENTIFY: typing.Final[int] = 2
@@ -328,15 +328,15 @@ class GatewayShardImpl(shard.GatewayShard):
 
     Parameters
     ----------
-    event_manager : hikari.api.event_manager.EventManager
-        The event manager this shard should make calls to.
-    event_factory : hikari.api.event_factory.EventFactory
-        The event factory this shard should use.
     token : builtins.str
         The bot token to use.
     url : builtins.str
         The gateway URL to use. This should not contain a query-string or
         fragments.
+    event_manager : hikari.api.event_manager.EventManager
+        The event manager this shard should make calls to.
+    event_factory : hikari.api.event_factory.EventFactory
+        The event factory this shard should use.
 
     Other Parameters
     ----------------
@@ -419,8 +419,6 @@ class GatewayShardImpl(shard.GatewayShard):
         self,
         *,
         compression: typing.Optional[str] = shard.GatewayCompression.PAYLOAD_ZLIB_STREAM,
-        event_factory: event_factory_.EventFactory,
-        event_manager: event_manager_.EventManager,
         initial_activity: typing.Optional[presences.Activity] = None,
         initial_idle_since: typing.Optional[datetime.datetime] = None,
         initial_is_afk: bool = False,
@@ -432,6 +430,8 @@ class GatewayShardImpl(shard.GatewayShard):
         http_settings: config.HTTPSettings,
         proxy_settings: config.ProxySettings,
         data_format: str = shard.GatewayDataFormat.JSON,
+        event_manager: event_manager_.EventManager,
+        event_factory: event_factory_.EventFactory,
         token: str,
         url: str,
     ) -> None:
@@ -457,8 +457,8 @@ class GatewayShardImpl(shard.GatewayShard):
             f"shard {shard_id} chunking rate limit",
             *_CHUNKING_RATELIMIT,
         )
-        self._event_factory = event_factory
         self._event_manager = event_manager
+        self._event_factory = event_factory
         self._handshake_completed = asyncio.Event()
         self._heartbeat_latency = float("nan")
         self._http_settings = http_settings
@@ -673,8 +673,7 @@ class GatewayShardImpl(shard.GatewayShard):
             self._handshake_completed.set()
 
         try:
-            self._event_manager.consume_raw_event(self, name, data)
-
+            self._event_manager.consume_raw_event(name, self, data)
         except LookupError:
             self._logger.debug("ignoring unknown event %s:\n    %r", name, data)
 
@@ -861,7 +860,7 @@ class GatewayShardImpl(shard.GatewayShard):
 
         try:
             # Dispatch CONNECTED synthetic event.
-            self._event_manager.dispatch(self._event_factory.deserialize_connected_event(self)),
+            self._event_manager.dispatch(self._event_factory.deserialize_connected_event(self))
             dispatch_disconnect = True
 
             heartbeat_task = await self._wait_for_hello()
@@ -931,7 +930,7 @@ class GatewayShardImpl(shard.GatewayShard):
             if dispatch_disconnect:
                 # If we managed to connect, we must always send the DISCONNECT event
                 # afterwards.
-                self._event_manager.dispatch(self._event_factory.deserialize_disconnected_event(self)),
+                self._event_manager.dispatch(self._event_factory.deserialize_disconnected_event(self))
 
             # Ignore errors if we are closing
             if self._closing.is_set() or self._closed.is_set():