… README.md
… hikari/internal/protocol.py
… hikari/traits.py
… pages/index.html
… scripts/protocol_benchmark.py
… tests/hikari/impl/test_event_manager_base.py
… tests/hikari/impl/test_shard.py
… tests/hikari/internal/test_protocol.py