- Remove unused args from slots
- Cleanup code
- Remove unused variables
- Fix documentation links

```
isinstance(long, BasicPyProtocol) 10.502764225999272 µs
isinstance(long, BasicHikariProtocol) 0.8962424719993578 µs
isinstance(failfast, BasicPyProtocol) 19.83291287300017 µs
isinstance(failfast, BasicHikariProtocol) 0.3696172330000991 µs
issubclass(long, BasicPyProtocol) 0.3248153929998807 µs
issubclass(long, BasicHikariProtocol) 0.30592563200116274 µs
issubclass(failfast, BasicPyProtocol) 0.3142782159993658 µs
issubclass(failfast, BasicHikariProtocol) 0.29908469400106696 µs
```

### Checklist
<!-- Make sure to tick all the following boxes by putting an `x` in between (like this `[x]`) -->
- [x] I have run `nox` and all the pipelines have passed.
- [x] I have made unittests according to the code I have added/modified/deleted.