@FasterSpeeding 2021-01-20
If the solution to avoiding instance checks against traits and try except duck checking in model methods then shouldn't shard aware and voice aware also be added to rest aware?