@FasterSpeeding 2021-01-20
Also doesn't this make the CacheAware trait useless as it can no longer be used to type something as needing a cache implementation aware client