@davfsa 2021-01-20
> If this is the final and best solution for avoiding instance checks against traits and try except duck checking in model methods then shouldn't shard aware and voice aware also be added to rest aware?

I don't see the reason to add them.