@FasterSpeeding 2021-01-20
> > If this is the final and best solution for avoiding instance checks against traits and try except duck checking in model methods then shouldn't shard aware and voice aware also be added to rest aware?
> 
> I don't see the reason to add them.

are you going to get rid of methods on entities that use shard aware if it's available and avoid adding voice methods to entities?